管理者画面のプロフィール画像編集を一部変更

+++ app/Http/Controllers/Administrator/cats/Update/PutController.php
@@ -16,7 +16,7 @@ class PutController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
-     */
+     */    
     public function __invoke(CatsRequest $request)
     {
         $cat = Cat::where('id', $request->id())->firstOrFail();
@@ -24,18 +24,31 @@ class PutController extends Controller
         DB::transaction( function () use($request, $cat) {
             $cat->name = $request->input('name');
             $cat->birthday = $request->input('birthday');
+            
+            $requestPhoto = $request->file('photo_URL');
 
             if ( !is_null($cat->photo_URL) ) {
-                Storage::disk('public')->delete('images/' . $cat->photo_URL);
+                if(!is_null($requestPhoto)){
+                    Storage::disk('public')->delete('images/' . $cat->photo_URL);
+                    $cat->photo_URL = $requestPhoto->hashName();
+                    $this->putImageToStorage($requestPhoto); 
+                }                
+            }else{
+                $cat->photo_URL = $requestPhoto->hashName();
+                $this->putImageToStorage($requestPhoto); 
             }
-    
-            $cat->photo_URL = $request->file('photo_URL')->hashName();
+            
             $cat->Instagram_URL = $request->input('Instagram_URL');
-            $cat->save();
-            Storage::putFile('public/images', $request->file('photo_URL'));                
+            $cat->save();               
         });
 
         return redirect()->route('administrator.cats.update.index', ['id' => $cat->id])
         ->with('feedback.success', "プロフィールを編集しました");
     }
+
+    private function putImageToStorage($image)
+    {
+        Storage::putFile('public/images', $image);
+    }
+
 }
